+++ .gitignore
@@ -1,3 +1,4 @@
 # Byte-compiled / optimized / DLL files
 __pycache__/
 data
+temp
